+++ Coroner/PluginConfig.cs
@@ -5,6 +5,7 @@ namespace Coroner
     public class PluginConfig
     {
         ConfigEntry<bool> DisplayCauseOfDeath;
+        ConfigEntry<bool> SeriousDeathMessages;
         ConfigEntry<bool> DisplayFunnyNotes;
         ConfigEntry<bool> DeathReplacesNotes;
 
@@ -17,6 +18,7 @@ namespace Coroner
         public void BindConfig(ConfigFile _config)
         {
             DisplayCauseOfDeath = _config.Bind("General", "DisplayCauseOfDeath", true, "Display the cause of death in the player notes.");
+            SeriousDeathMessages = _config.Bind("General", "SeriousDeathMessages", false, "Cause of death messages are more to-the-point.");
             DisplayFunnyNotes = _config.Bind("General", "DisplayFunnyNotes", true, "Display a random note when the player has no notes.");
             DeathReplacesNotes = _config.Bind("General", "DeathReplacesNotes", true, "True to replace notes when the player dies, false to append.");
         }
@@ -24,6 +26,11 @@ namespace Coroner
         public bool ShouldDisplayCauseOfDeath()
         {
             return DisplayCauseOfDeath.Value;
+        }
+
+        public bool ShouldUseSeriousDeathMessages()
+        {
+            return SeriousDeathMessages.Value;
         }
 
         public bool ShouldDisplayFunnyNotes()